Enhance type safety and improve event handling
- Updated `IdHandler` to explicitly cast bitwise operation results to `int`.
- Introduced a new `Event` enum in `Event.cs` for better event type management.
- Modified `HandshakeResponse` to adjust `totalExtraData` calculation and ensure complete data reading from the stream.
- Refactored `Screen` class to use a private field for `RootDepth` and simplified the `FillTheDepth` method for better performance.
- Improved window creation request handling in `Program.cs` by using stack-allocated spans and refactoring event sending logic for enhanced clarity and error handling.

+++ src/Handler/IdHandler.cs
@@ -11,9 +11,9 @@ namespace src.Handler
         private static int _globalId = 0;
         public static int GetId(HandshakeResponseSuccessBody handshakeResponseSuccess)
         {
-            var result = (handshakeResponseSuccess.ResourceIDMask & _globalId) | handshakeResponseSuccess.ResourceIDBase;
+            var result = (int)((handshakeResponseSuccess.ResourceIDMask & _globalId) | handshakeResponseSuccess.ResourceIDBase);
             _globalId += 1;
-            return (int)result;
+            return result;
         }
     }
 }
